+++ src/helpers/dropdown.ts
@@ -1,5 +1,4 @@
 import { NormalizedDropdownItemType } from '@entur/dropdown';
-import { FormatMessage } from 'i18n';
 import { FlexibleLineType } from 'model/FlexibleLine';
 import { MessagesKey } from '../i18n/translations/translationKeys';
 import {
@@ -7,6 +6,7 @@ import {
   VEHICLE_SUBMODE,
   vehicleSubmodeMessages,
 } from '../model/enums';
+import { FormatMessage } from 'i18n';
 
 type GenericDropdownItem<E> = {
   id?: string;
