+++ src/i18n/intlSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import type { RootState } from 'store/store';
-import { getLocale, LOCALE_KEY } from './index';
+import { getLocale, LOCALE_KEY } from './getLocale';
 import { Locale } from './locale';
 
 export interface LocaleState {
