+++ src/scenes/App/components/NavBar/LanguagePicker/index.tsx
@@ -1,47 +1,15 @@
 import { FloatingButton } from '@entur/button';
-import {
-  CheckIcon,
-  DownArrowIcon,
-  NorwayIcon,
-  SwedenIcon,
-  UKIcon,
-  UpArrowIcon,
-} from '@entur/icons';
+import { CheckIcon, DownArrowIcon, UpArrowIcon } from '@entur/icons';
 import { updateLocale } from 'i18n/intlSlice';
-import { MessagesKey } from 'i18n/translations/translationKeys';
 import { useCallback, useMemo, useState } from 'react';
 import { useIntl } from 'react-intl';
 import { useDispatch } from 'react-redux';
-import './styles.scss';
-import { useConfig } from '../../../../../config/ConfigContext';
-import { Locale, SUPPORTED_LOCALES } from '../../../../../i18n/locale';
-
-const getFlagIcon = (locale: string) => {
-  switch (locale) {
-    case 'nb':
-      return <NorwayIcon inline />;
-    case 'sv':
-      return <SwedenIcon inline />;
-    case 'en':
-      return <UKIcon inline />;
-    default: {
-      return <NorwayIcon inline />;
-    }
-  }
-};
-
-const getLocaleString = (locale: string): keyof MessagesKey => {
-  switch (locale) {
-    case 'en':
-      return 'userMenuMenuItemTextEnglish';
-    case 'sv':
-      return 'userMenuMenuItemTextSwedish';
-    case 'nb':
-      return 'userMenuMenuItemTextNorwegian';
-    default:
-      return 'userMenuMenuItemTextNorwegian';
-  }
-};
+import { useConfig } from 'config/ConfigContext';
+import {
+  getLanguagePickerFlagIcon,
+  getLanguagePickerLocaleMessageKey,
+  Locale,
+} from 'i18n';
 
 const LanguagePicker = () => {
   const [toggled, setToggle] = useState<boolean>(false);
@@ -77,10 +45,10 @@ const LanguagePicker = () => {
     [selectedLocale],
   );
 
-  const flagIcon = useCallback((locale: string) => {
+  const flagIcon = useCallback((locale: Locale) => {
     return (
       <div className="language-picker__icon-container language-picker__icon-container__flag">
-        {getFlagIcon(locale)}
+        {getLanguagePickerFlagIcon(locale)}
       </div>
     );
   }, []);
@@ -89,7 +57,7 @@ const LanguagePicker = () => {
     <div className="language-picker-wrapper">
       {toggled && (
         <div className="language-picker__dropdown">
-          {(supportedLocales || SUPPORTED_LOCALES).map((locale: Locale) => (
+          {(supportedLocales || Locale).map((locale: Locale) => (
             <FloatingButton
               key={locale}
               className="language-picker__item"
@@ -98,7 +66,11 @@ const LanguagePicker = () => {
               size="small"
             >
               {flagIcon(locale)}
-              <span>{formatMessage({ id: getLocaleString(locale) })}</span>
+              <span>
+                {formatMessage({
+                  id: getLanguagePickerLocaleMessageKey(locale),
+                })}
+              </span>
               {checkIcon(locale)}
             </FloatingButton>
           ))}
@@ -110,8 +82,10 @@ const LanguagePicker = () => {
         className="language-picker"
         size="small"
       >
-        {flagIcon(selectedLocale)}
-        {formatMessage({ id: getLocaleString(selectedLocale) })}
+        {flagIcon(selectedLocale as Locale)}
+        {formatMessage({
+          id: getLanguagePickerLocaleMessageKey(selectedLocale as Locale),
+        })}
         {arrowIcon}
       </FloatingButton>
     </div>
